finally got the last image module working.  Installed older version to get it to work
