+++ package.json
@@ -12,8 +12,11 @@
     "author": "",
     "license": "ISC",
     "devDependencies": {
+        "file-loader": "^6.2.0",
+        "image-webpack-loader": "^6.0.0",
         "webpack": "^5.64.2",
-        "webpack-cli": "^3.3.11"
+        "webpack-bundle-analyzer": "^4.5.0",
+        "webpack-cli": "^4.9.1"
     },
     "dependencies": {
         "bootstrap": "^4.3.1",
