+++ schedule.html
@@ -9,7 +9,7 @@
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
   <link rel="stylesheet" href="./assets/css/style.css">
 
-  <script src="./dist/main.bundle.js" defer ></script>
+  <script src="./dist/schedule.bundle.js" defer ></script>
 </head>
 
 <body class="schedule-bg">
