+++ tickets.html
@@ -10,7 +10,7 @@
   <link rel="stylesheet" href="./assets/css/tickets.css">
   <link rel="stylesheet" href="./assets/css/style.css">
 
-  <script src="./dist/main.bundle.js" defer ></script>
+  <script src="./dist/tickets.bundle.js" defer ></script>
 </head>
 <body class="tickets-bg">
   <nav class="navbar navbar-expand-lg navbar-light bg-light fixed-top">
