+++ webpack.config.js
@@ -1,17 +1,53 @@
-const webpack = require("webpack")
+const webpack = require("webpack");
+const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
+
 const path = require("path");
 
-module.exports = {
-  entry: "./assets/js/script.js",
+const config = {
+  entry: {
+    app: "./assets/js/script.js",
+    events: "./assets/js/events.js",
+    schedule: "./assets/js/schedule.js",
+    tickets: "./assets/js/tickets.js"
+    },
   output: {
-    path: path.join(__dirname + "/dist"),
-    filename: "main.bundle.js"
+    filename: "[name].bundle.js",
+    path: __dirname + "/dist",
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(png|jpe?g|gif)$/i,
+        use: [
+          {
+            loader: 'file-loader',
+            options: {
+              esModule: false,
+              name(file) {
+                return '[path][name].[ext]';
+              },
+              publicPath: function(url) {
+                return url.replace('../', '/assets/');
+              }
+            }
+          },
+          {
+            loader: 'image-webpack-loader'
+          }
+        ]
+      }
+    ]
   },
   plugins: [
   new webpack.ProvidePlugin({
     $: "jquery",
     jQuery: "jquery"
   }),
+  new BundleAnalyzerPlugin({
+      analyzerMode: "static", // the report outputs to an HTML file in the dist folder
+  })
   ],
   mode: "development"
 };
+
+module.exports = config;
